Fix number pattern of `stringify` function to support exponential log value.
When `typia.stringify<T>()` be used for an object type which has template typed key containing a number type, and exponential log being used in the template key type, `typia` could not write the property because regex pattern for numeric value of ` typia` had not supported the exponential log value.

This is an extremely rare case, but fixed for extensionality.

```typescript
export interface ISomething {
    [key: `value_${number}`]: boolean | string | number;
}

const something: ISomething = {
    "5.175933557310941e-7": -0.170261004873707,
};
typia.stringify(something);
```