Change convert of comment tag to type tag.
When converting comment tag to type name, its representative name was a little bit wrong when `Format` or `Pattern` category case. It must be `Format<"uuid">`, but had converted to `Format<uuid>`. It had not affected to validation logic and had not caused any bug, but I've fixed it to exact expression.