Enhance Yarn Berry Comment
This comment is misleading and may lead to users assuming it is not possible to use yarn berry at all.  

Instead, link to the github and indicate it can stll be used manually - at the moment this is not linked in docs that i could find and i almost left the site!