Fix #1001: constant union type in Array.
When constant union type being used in an Array, `typia.protobuf` functions had thrown compilation error with "union type is not allowed in an Array type".

By the way, constant union type like `Array<1|2|3>` must be allowed. This PR fixes the bug, so that `Array<1|2|3>` type would be supported from now on.