Fix bug of `application_plugin()` function.
When type tags being utilized, but none of them has the `schema` property, `typia.json.application<T>()` failed to generate the JSON schema with "invalid type on argument" error message.

The bug has occured by mistake on the filter implementation of `application_plugin()` function. This PR fixes the bug, so that same error would not be occured again.