feat(workflows/build.yaml): split npm/bun jobs
This commit introduces a new GitHub Actions workflow for the Ubuntu
platform. It sets up Node.js, caches npm dependencies, and runs the
build and test scripts. Additionally, it adds a new job for the
actions-timeline action, which visualizes the workflow execution
timeline. The existing Ubuntu-Bun job is also updated to remove the
--skipSetup flag from the test:bun script.