feat: Improve command line argument parsing
This commit introduces the use of `parseArgs` from `node:util` to handle
command line arguments in a more structured and robust way. This change
affects the `deploy/bun.ts` and `deploy/index.ts` files.

In addition, the scripts in `package.json` have been updated to reflect
the new way of passing arguments to the deploy scripts.