feat: Add Build-and-Upload-topkg-pr-new job to workflow
This commit introduces a new job in the GitHub Actions workflow. The
job, named Build-and-Upload-topkg-pr-new, is designed to run on the
latest Ubuntu, set up Node.js version 20.x, cache npm dependencies,
install them, build the project, and publish using pkg-pr-new. This
job is added as a dependency for the action-timeline job.