feat(build.yml): Update build and upload steps
This commit updates the GitHub Actions workflow file 'build.yml'. The
'Build-and-Upload-topkg-pr-new' job is renamed to
'Build-and-Upload-to-pkg-pr-new-and-artifactory'. The step to upload
the build artifact to Artifactory is moved from the 'NodeJS' job to
the renamed job. This change ensures that the artifact is uploaded
only once and in the correct job.