feat(utils): add findTsConfig function
This commit introduces a new utility function `findTsConfig`. This
function is used to find the `tsconfig.json` file in the project. If
the `project` argument is provided, it returns it. If not, it tries
to find the `tsconfig.json` file in the directory hierarchy. If the
file is not found, it throws an error.

The `generateCommand` in `TypiaGenerateWizard.ts` is also updated to
use this new function, simplifying its code.